Refactor: Consolidate backtesting engine logic
Move backtesting engine implementation to a central location within src/utils/nas_tas.
