+++ src/utils/nas_tas/__init__.py
@@ -1,71 +1,20 @@
 """
-Unified NAS/TAS Utilities
+NAS TAS (Neural Architecture Search - Tree Architecture Search) Utilities
 
-This module provides unified utilities for both Neural Architecture Search (NAS)
-and Tree Architecture Search (TAS) systems.
+This module provides common utilities for NAS and TAS operations including
+backtesting engines, regime detection, and optimization tools.
 """
 
-from .unified_evaluator import (
-    UnifiedEvaluator,
-    EvaluationConfig,
-    EvaluationResult,
-    ModelType,
-    EvaluationMode,
-    MetricType
-)
-
-from .unified_multi_objective import (
-    UnifiedMultiObjectiveOptimizer,
-    PerformanceEstimator,
-    ArchitectureFeatures,
-    PerformancePrediction,
-    PerformanceMetric,
-    EstimatorType,
-    OptimizationConfig,
-    MultiObjectiveResult
-)
-
-from .bayesian_search import (
-    BayesianTreeSearch,
-    TreeBayesianOptimizer,
-    TreeGaussianProcess,
-    BayesianConfig
-)
-
-from .search_space import (
-    SearchSpace,
-    SearchSpaceConfig,
-    ParameterRange,
-    SearchSpaceType,
-    OptimizationStrategy,
-    create_default_nas_search_space,
-    create_tree_search_space
+from .backtesting_engine import (
+    BacktestingEngine,
+    BacktestingConfig,
+    BacktestingResult,
+    BacktestingMode
 )
 
 __all__ = [
-    'UnifiedEvaluator',
-    'EvaluationConfig', 
-    'EvaluationResult',
-    'ModelType',
-    'EvaluationMode',
-    'MetricType',
-    'UnifiedMultiObjectiveOptimizer',
-    'PerformanceEstimator',
-    'ArchitectureFeatures',
-    'PerformancePrediction',
-    'PerformanceMetric',
-    'EstimatorType',
-    'OptimizationConfig',
-    'MultiObjectiveResult',
-    'BayesianTreeSearch',
-    'TreeBayesianOptimizer',
-    'TreeGaussianProcess',
-    'BayesianConfig',
-    'SearchSpace',
-    'SearchSpaceConfig',
-    'ParameterRange',
-    'SearchSpaceType',
-    'OptimizationStrategy',
-    'create_default_nas_search_space',
-    'create_tree_search_space'
+    'BacktestingEngine',
+    'BacktestingConfig', 
+    'BacktestingResult',
+    'BacktestingMode'
 ]
